+++ node.gyp
@@ -345,6 +345,7 @@
         'src/node_domain.cc',
         'src/node_encoding.cc',
         'src/node_errors.h',
+        'src/node_errors.cc',
         'src/node_file.cc',
         'src/node_http2.cc',
         'src/node_http_parser.cc',
