+++ src/async_wrap.cc
@@ -21,9 +21,10 @@
 
 #include "async_wrap-inl.h"
 #include "env-inl.h"
+#include "node_errors.h"
 #include "node_internals.h"
-#include "util-inl.h"
 #include "tracing/traced_value.h"
+#include "util-inl.h"
 
 #include "v8.h"
 #include "v8-profiler.h"
