+++ src/inspector_agent.cc
@@ -1,12 +1,13 @@
 #include "inspector_agent.h"
 
-#include "inspector_io.h"
 #include "inspector/main_thread_interface.h"
 #include "inspector/node_string.h"
 #include "inspector/tracing_agent.h"
 #include "inspector/worker_agent.h"
 #include "inspector/worker_inspector.h"
+#include "inspector_io.h"
 #include "node/inspector/protocol/Protocol.h"
+#include "node_errors.h"
 #include "node_internals.h"
 #include "node_url.h"
 #include "v8-inspector.h"
