+++ src/node_url.cc
@@ -1,7 +1,8 @@
 #include "node_url.h"
-#include "node_internals.h"
 #include "base_object-inl.h"
+#include "node_errors.h"
 #include "node_i18n.h"
+#include "node_internals.h"
 
 #include <string>
 #include <vector>
